requirements.txt
psycopg2-binary
fastapi[standard]
uvicorn[standard]
sqlalchemy
pydantic
alembic
python-dotenv
httpx
pytest
psycopg2
databases[postgresql]
pytest-asyncio
rapidfuzz
pytest-asyncio
bcrypt
email-validator
PyJWT
python-multipart
passlib[bcrypt]
python-decouple
redis
botocore
boto3
celery
decouple